requirements.txt
psycopg2-binary==2.9.9
pyyaml==6.0.1
google-cloud-sql==0.3.0
python-dotenv==1.0.0
google-cloud-scheduler==2.12.0 